Closes pablofalegre/pdes-tips#8 . Se separaron las rutas en distintos archivos.
